+++ src/main/java/com/acuo/persist/services/MarginStatementServiceImpl.java
@@ -171,10 +171,10 @@ public class MarginStatementServiceImpl extends GenericService<MarginStatement, 
 
     @Override
     @Transactional
-    public int getCountForMenu(String status)
+    public Long getCountForMenu(String status)
     {
         String query = "MATCH (ms:MarginStatement )<-[:PART_OF]-(s:StatementItem)-[:LAST]->(step:Step {status:{status}}) " +
                 "RETURN count(distinct(ms)) AS count;";
-        return (Integer) sessionProvider.get().query(query, ImmutableMap.of("status", status)).iterator().next().get("count");
+        return (Long) sessionProvider.get().query(query, ImmutableMap.of("status", status)).iterator().next().get("count");
     }
 }
